Refactoring page name; replacing the home page with the summary page
